dultdashboard/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
gunicorn==20.1.0